backend/requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
mongoengine==0.28.2
python-dotenv==1.0.0
bcrypt==4.1.2
google-generativeai==0.3.2
openai==1.6.1
boto3==1.34.14
Pillow==10.2.0
gunicorn==21.2.0
